fix: incorrect file extension
